+++ muic_v2/GalleryDetailController.m
@@ -19,7 +19,7 @@
 
 @synthesize cvMain,contentList;
 
-int selectedItemIndex;
+NSInteger selectedItemIndex;
 
 - (void)viewDidLoad
 {
